feat(parser): validate type annotations against inferred types
Add annotation consistency checking to the parser. When a variable has
a type annotation, the parser now validates it against the inferred type
and reports mismatches (shape, dtype, rank, kind) with actionable errors.

- Add validate_annotation_consistency to TypeResolver
- Add _dtypes_compatible helper (INDEX ↔ integer compatibility)
- Remove dead resolve_type_if_memref method
- Fix incorrect annotations in tests to match actual inferred types

+++ tests/ut/language/parser/test_text_parser.py
@@ -237,8 +237,8 @@ def complex_op(
 ) -> pl.Tensor[[64, 128], pl.FP16]:
     # Multiple operations
     temp1: pl.Tensor[[64, 128], pl.FP16] = pl.add(x, y)
-    temp2: pl.Tensor[[64, 128], pl.FP16] = pl.mul(temp1, 2.0)
-    result: pl.Tensor[[64, 128], pl.FP16] = pl.sub(temp2, x)
+    temp2: pl.Tensor[[64, 128], pl.FP32] = pl.mul(temp1, 2.0)
+    result: pl.Tensor[[64, 128], pl.FP32] = pl.sub(temp2, x)
     return result
 """
         with tempfile.NamedTemporaryFile(mode="w", suffix=".py", delete=False) as f:
